refactor: refactor generateText to use useGenerationExecutor
- Reduce code duplication by using shared generation executor pattern
- Extract common logic for generation setup, usage limits, and error handling
- Simplify generateText to focus on text generation specific logic
- Maintain type safety and existing functionality
- Reduce file size from 442 to 287 lines (~35% reduction)
