+++ packages/giselle-engine/src/core/generations/generate-text.ts
@@ -5,18 +5,10 @@ import { perplexity } from "@ai-sdk/perplexity";
 import {
 	type CompletedGeneration,
 	type FailedGeneration,
-	type FileData,
-	GenerationContext,
 	type GenerationOutput,
-	type NodeId,
-	type Output,
-	type OutputId,
 	type QueuedGeneration,
-	type RunningGeneration,
 	type TextGenerationLanguageModelData,
 	type UrlSource,
-	type WorkspaceId,
-	isCompletedGeneration,
 	isTextGenerationNode,
 } from "@giselle-sdk/data-type";
 import { githubTools, octokit } from "@giselle-sdk/github-tool";
@@ -27,21 +19,12 @@ import {
 	languageModels,
 } from "@giselle-sdk/language-model";
 import { AISDKError, appendResponseMessages, streamText } from "ai";
-import { UsageLimitError } from "../error";
-import { filePath } from "../files/utils";
 import type { GiselleEngineContext } from "../types";
-import { internalSetGeneration } from "./internal/set-generation";
+import { useGenerationExecutor } from "./internal/use-generation-executor";
 import { createLangfuseTracer, generateTelemetryTags } from "./telemetry";
 import { createPostgresTools } from "./tools/postgres";
 import type { PreparedToolSet, TelemetrySettings } from "./types";
-import {
-	buildMessageObject,
-	checkUsageLimits,
-	getGeneration,
-	getNodeGenerationIndexes,
-	handleAgentTimeConsumption,
-	queryResultToText,
-} from "./utils";
+import { buildMessageObject, handleAgentTimeConsumption } from "./utils";
 
 // PerplexityProviderOptions is not exported from @ai-sdk/perplexity, so we define it here based on the model configuration
 export type PerplexityProviderOptions = {
@@ -53,442 +36,287 @@ export async function generateText(args: {
 	generation: QueuedGeneration;
 	telemetry?: TelemetrySettings;
 }) {
-	const operationNode = args.generation.context.operationNode;
-	if (!isTextGenerationNode(operationNode)) {
-		throw new Error("Invalid generation type");
-	}
-
-	const languageModel = languageModels.find(
-		(lm) => lm.id === operationNode.content.llm.id,
-	);
-	if (!languageModel) {
-		throw new Error("Invalid language model");
-	}
-	const runningGeneration = {
-		...args.generation,
-		status: "running",
-		messages: [],
-		startedAt: Date.now(),
-	} satisfies RunningGeneration;
-
-	await internalSetGeneration({
-		storage: args.context.storage,
-		generation: runningGeneration,
-	});
-
-	let workspaceId: WorkspaceId | undefined;
-	switch (args.generation.context.origin.type) {
-		case "run":
-			workspaceId = args.generation.context.origin.workspaceId;
-			break;
-		case "workspace":
-			workspaceId = args.generation.context.origin.id;
-			break;
-		default: {
-			const _exhaustiveCheck: never = args.generation.context.origin;
-			throw new Error(`Unhandled origin type: ${_exhaustiveCheck}`);
-		}
-	}
-
-	const usageLimitStatus = await checkUsageLimits({
-		workspaceId,
+	return useGenerationExecutor({
+		context: args.context,
 		generation: args.generation,
-		fetchUsageLimitsFn: args.context.fetchUsageLimitsFn,
-	});
-	if (usageLimitStatus.type === "error") {
-		const failedGeneration = {
-			...runningGeneration,
-			status: "failed",
-			failedAt: Date.now(),
-			error: {
-				name: usageLimitStatus.error,
-				message: usageLimitStatus.error,
-				dump: usageLimitStatus,
-			},
-		} satisfies FailedGeneration;
+		execute: async ({
+			runningGeneration,
+			generationContext,
+			setGeneration,
+			fileResolver,
+			generationContentResolver,
+			workspaceId,
+		}) => {
+			const operationNode = generationContext.operationNode;
+			if (!isTextGenerationNode(operationNode)) {
+				throw new Error("Invalid generation type");
+			}
 
-		await internalSetGeneration({
-			storage: args.context.storage,
-			generation: failedGeneration,
-		});
+			const languageModel = languageModels.find(
+				(lm) => lm.id === operationNode.content.llm.id,
+			);
+			if (!languageModel) {
+				throw new Error("Invalid language model");
+			}
 
-		throw new UsageLimitError(usageLimitStatus.error);
-	}
+			const messages = await buildMessageObject(
+				operationNode,
+				runningGeneration.context.sourceNodes,
+				fileResolver,
+				generationContentResolver,
+			);
 
-	async function fileResolver(file: FileData) {
-		const blob = await args.context.storage.getItemRaw(
-			filePath({
-				...runningGeneration.context.origin,
-				fileId: file.id,
-			}),
-		);
-		if (blob === undefined) {
-			return undefined;
-		}
-		return blob;
-	}
+			let preparedToolSet: PreparedToolSet = {
+				toolSet: {},
+				cleanupFunctions: [],
+			};
 
-	async function generationContentResolver(nodeId: NodeId, outputId: OutputId) {
-		const nodeGenerationIndexes = await getNodeGenerationIndexes({
-			storage: args.context.storage,
-			nodeId,
-		});
-		if (
-			nodeGenerationIndexes === undefined ||
-			nodeGenerationIndexes.length === 0
-		) {
-			return undefined;
-		}
-		const generation = await getGeneration({
-			...args,
-			storage: args.context.storage,
-			generationId: nodeGenerationIndexes[nodeGenerationIndexes.length - 1].id,
-			options: {
-				bypassingCache: true,
-			},
-		});
-		if (generation === undefined || !isCompletedGeneration(generation)) {
-			return undefined;
-		}
-		let output: Output | undefined;
-		for (const sourceNode of runningGeneration.context.sourceNodes) {
-			for (const sourceOutput of sourceNode.outputs) {
-				if (sourceOutput.id === outputId) {
-					output = sourceOutput;
-					break;
+			if (operationNode.content.tools?.github?.auth) {
+				const decryptToken = await args.context.vault?.decrypt(
+					operationNode.content.tools.github.auth.token,
+				);
+				const allGitHubTools = githubTools(
+					octokit({
+						strategy: "personal-access-token",
+						personalAccessToken:
+							decryptToken ?? operationNode.content.tools.github.auth.token,
+					}),
+				);
+				for (const tool of operationNode.content.tools.github.tools) {
+					if (tool in allGitHubTools) {
+						preparedToolSet = {
+							...preparedToolSet,
+							toolSet: {
+								...preparedToolSet.toolSet,
+								[tool]: allGitHubTools[tool as keyof typeof allGitHubTools],
+							},
+						};
+					}
 				}
 			}
-		}
-		if (output === undefined) {
-			return undefined;
-		}
-		const generationOutput = generation.outputs.find(
-			(output) => output.outputId === outputId,
-		);
-		if (generationOutput === undefined) {
-			return undefined;
-		}
-		switch (generationOutput.type) {
-			case "source":
-				return JSON.stringify(generationOutput.sources);
-			case "reasoning":
-				throw new Error("Generation output type is not supported");
-			case "generated-image":
-				throw new Error("Generation output type is not supported");
-			case "generated-text":
-				return generationOutput.content;
-			case "query-result": {
-				return queryResultToText(generationOutput);
-			}
-			default: {
-				const _exhaustiveCheck: never = generationOutput;
-				throw new Error(
-					`Unhandled generation output type: ${_exhaustiveCheck}`,
-				);
-			}
-		}
-	}
-	const messages = await buildMessageObject(
-		operationNode,
-		runningGeneration.context.sourceNodes,
-		fileResolver,
-		generationContentResolver,
-	);
 
-	let preparedToolSet: PreparedToolSet = { toolSet: {}, cleanupFunctions: [] };
-	if (operationNode.content.tools?.github?.auth) {
-		const decryptToken = await args.context.vault?.decrypt(
-			operationNode.content.tools.github.auth.token,
-		);
-		const allGitHubTools = githubTools(
-			octokit({
-				strategy: "personal-access-token",
-				personalAccessToken:
-					decryptToken ?? operationNode.content.tools.github.auth.token,
-			}),
-		);
-		for (const tool of operationNode.content.tools.github.tools) {
-			if (tool in allGitHubTools) {
+			if (operationNode.content.tools?.postgres?.connectionString) {
+				const connectionString = await args.context.vault?.decrypt(
+					operationNode.content.tools.postgres.connectionString,
+				);
+				const postgresTool = createPostgresTools(
+					connectionString ??
+						operationNode.content.tools.postgres.connectionString,
+				);
+				for (const tool of operationNode.content.tools.postgres.tools) {
+					if (tool in postgresTool.toolSet) {
+						preparedToolSet = {
+							...preparedToolSet,
+							toolSet: {
+								...preparedToolSet.toolSet,
+								[tool]:
+									postgresTool.toolSet[
+										tool as keyof typeof postgresTool.toolSet
+									],
+							},
+						};
+					}
+					preparedToolSet = {
+						...preparedToolSet,
+						cleanupFunctions: [
+							...preparedToolSet.cleanupFunctions,
+							postgresTool.cleanup,
+						],
+					};
+				}
+			}
+
+			if (
+				operationNode.content.llm.provider === "openai" &&
+				operationNode.content.tools?.openaiWebSearch &&
+				hasCapability(languageModel, Capability.OptionalSearchGrounding)
+			) {
 				preparedToolSet = {
 					...preparedToolSet,
 					toolSet: {
 						...preparedToolSet.toolSet,
-						[tool]: allGitHubTools[tool as keyof typeof allGitHubTools],
+						openaiWebSearch: openai.tools.webSearchPreview(
+							operationNode.content.tools.openaiWebSearch,
+						),
 					},
 				};
 			}
-		}
-	}
 
-	if (operationNode.content.tools?.postgres?.connectionString) {
-		const connectionString = await args.context.vault?.decrypt(
-			operationNode.content.tools.postgres.connectionString,
-		);
-		const postgresTool = createPostgresTools(
-			connectionString ?? operationNode.content.tools.postgres.connectionString,
-		);
-		for (const tool of operationNode.content.tools.postgres.tools) {
-			if (tool in postgresTool.toolSet) {
-				preparedToolSet = {
-					...preparedToolSet,
-					toolSet: {
-						...preparedToolSet.toolSet,
-						[tool]:
-							postgresTool.toolSet[tool as keyof typeof postgresTool.toolSet],
-					},
-				};
-			}
-			preparedToolSet = {
-				...preparedToolSet,
-				cleanupFunctions: [
-					...preparedToolSet.cleanupFunctions,
-					postgresTool.cleanup,
-				],
-			};
-		}
-	}
+			const providerOptions = getProviderOptions(operationNode.content.llm);
 
-	if (
-		operationNode.content.llm.provider === "openai" &&
-		operationNode.content.tools?.openaiWebSearch &&
-		hasCapability(languageModel, Capability.OptionalSearchGrounding)
-	)
-		preparedToolSet = {
-			...preparedToolSet,
-			toolSet: {
-				...preparedToolSet.toolSet,
-				openaiWebSearch: openai.tools.webSearchPreview(
-					operationNode.content.tools.openaiWebSearch,
-				),
-			},
-		};
+			const streamTextResult = streamText({
+				model: generationModel(operationNode.content.llm),
+				providerOptions,
+				messages,
+				maxSteps: 5, // enable multi-step calls
+				tools: preparedToolSet.toolSet,
+				experimental_continueSteps: true,
+				onError: async ({ error }) => {
+					if (AISDKError.isInstance(error)) {
+						const failedGeneration = {
+							...runningGeneration,
+							status: "failed",
+							failedAt: Date.now(),
+							error: {
+								name: error.name,
+								message: error.message,
+							},
+						} satisfies FailedGeneration;
 
-	// if (
-	// 	operationNode.content.tools?.github &&
-	// 	args.context.integrationConfigs?.github
-	// ) {
-	// 	const auth = args.context.integrationConfigs.github.auth;
-	// 	switch (auth.strategy) {
-	// 		case "app-installation": {
-	// 			const installationId = await auth.resolver.installationIdForRepo(
-	// 				operationNode.content.tools.github.repositoryNodeId,
-	// 			);
-	// 			const allGitHubTools = githubTools(
-	// 				octokit({
-	// 					...auth,
-	// 					installationId,
-	// 				}),
-	// 			);
-	// 			for (const tool of operationNode.content.tools.github.tools) {
-	// 				if (tool in allGitHubTools) {
-	// 					tools = {
-	// 						...tools,
-	// 						[tool]: allGitHubTools[tool as keyof typeof allGitHubTools],
-	// 					};
-	// 				}
-	// 			}
-	// 			break;
-	// 		}
-	// 		case "personal-access-token": {
-	// 			const allGitHubTools = githubTools(octokit(auth));
-	// 			for (const tool of operationNode.content.tools.github.tools) {
-	// 				if (tool in allGitHubTools) {
-	// 					tools = {
-	// 						...tools,
-	// 						[tool]: allGitHubTools[tool as keyof typeof allGitHubTools],
-	// 					};
-	// 				}
-	// 			}
-	// 			break;
-	// 		}
-	// 		default: {
-	// 			const _exhaustiveCheck: never = auth;
-	// 			throw new Error(`Unhandled GitHub auth strategy: ${_exhaustiveCheck}`);
-	// 		}
-	// 	}
-	// }
+						await setGeneration(failedGeneration);
+					}
 
-	const providerOptions = getProviderOptions(operationNode.content.llm);
-
-	const streamTextResult = streamText({
-		model: generationModel(operationNode.content.llm),
-		providerOptions,
-		messages,
-		maxSteps: 5, // enable multi-step calls
-		tools: preparedToolSet.toolSet,
-		experimental_continueSteps: true,
-		onError: async ({ error }) => {
-			if (AISDKError.isInstance(error)) {
-				const failedGeneration = {
-					...runningGeneration,
-					status: "failed",
-					failedAt: Date.now(),
-					error: {
-						name: error.name,
-						message: error.message,
-					},
-				} satisfies FailedGeneration;
-
-				await internalSetGeneration({
-					storage: args.context.storage,
-					generation: failedGeneration,
-				});
-			}
-
-			await Promise.all(
-				preparedToolSet.cleanupFunctions.map((cleanupFunction) =>
-					cleanupFunction(),
-				),
-			);
-		},
-		async onFinish(event) {
-			const generationContext = GenerationContext.parse(
-				args.generation.context,
-			);
-			const generationOutputs: GenerationOutput[] = [];
-			const generatedTextOutput = generationContext.operationNode.outputs.find(
-				(output) => output.accessor === "generated-text",
-			);
-			if (generatedTextOutput !== undefined) {
-				generationOutputs.push({
-					type: "generated-text",
-					content: event.text,
-					outputId: generatedTextOutput.id,
-				});
-			}
-
-			const tokenUsage = event.usage;
-			let costInfo = null;
-
-			if (tokenUsage) {
-				costInfo = await calculateDisplayCost(
-					operationNode.content.llm.provider,
-					operationNode.content.llm.id,
-					tokenUsage,
-				);
-			}
-
-			const reasoningOutput = generationContext.operationNode.outputs.find(
-				(output) => output.accessor === "reasoning",
-			);
-			if (reasoningOutput !== undefined && event.reasoning !== undefined) {
-				generationOutputs.push({
-					type: "reasoning",
-					content: event.reasoning,
-					outputId: reasoningOutput.id,
-				});
-			}
-			const sourceOutput = generationContext.operationNode.outputs.find(
-				(output) => output.accessor === "source",
-			);
-			if (sourceOutput !== undefined && event.sources.length > 0) {
-				const sources = await Promise.all(
-					event.sources.map(async (source) => {
-						return {
-							sourceType: "url",
-							id: source.id,
-							url: source.url,
-							title: source.title ?? source.url,
-							providerMetadata: source.providerMetadata,
-						} satisfies UrlSource;
-					}),
-				);
-				generationOutputs.push({
-					type: "source",
-					outputId: sourceOutput.id,
-					sources,
-				});
-			}
-			const completedGeneration = {
-				...runningGeneration,
-				status: "completed",
-				completedAt: Date.now(),
-				outputs: generationOutputs,
-				usage: tokenUsage,
-				messages: appendResponseMessages({
-					messages: [
-						{
-							id: "id",
-							role: "user",
-							content: "",
-						},
-					],
-					responseMessages: event.response.messages,
-				}),
-			} satisfies CompletedGeneration;
-
-			await internalSetGeneration({
-				storage: args.context.storage,
-				generation: completedGeneration,
-			});
-
-			await handleAgentTimeConsumption({
-				workspaceId,
-				generation: completedGeneration,
-				onConsumeAgentTime: args.context.onConsumeAgentTime,
-			});
-
-			// necessary to send telemetry but not explicitly used
-			const langfuse = createLangfuseTracer({
-				workspaceId,
-				runningGeneration,
-				tags: generateTelemetryTags({
-					provider: operationNode.content.llm.provider,
-					languageModel,
-					toolSet: preparedToolSet.toolSet,
-					configurations: operationNode.content.llm.configurations,
-					providerOptions,
-				}),
-				messages: { messages },
-				output: event.text,
-				usage: {
-					input: tokenUsage?.promptTokens ?? 0,
-					output: tokenUsage?.completionTokens ?? 0,
-					total:
-						(tokenUsage?.promptTokens ?? 0) +
-						(tokenUsage?.completionTokens ?? 0),
-					inputCost: costInfo?.inputCostForDisplay ?? 0,
-					outputCost: costInfo?.outputCostForDisplay ?? 0,
-					totalCost: costInfo?.totalCostForDisplay ?? 0,
-					unit: "TOKENS",
+					await Promise.all(
+						preparedToolSet.cleanupFunctions.map((cleanupFunction) =>
+							cleanupFunction(),
+						),
+					);
 				},
-				completedGeneration,
-				spanName: "ai.streamText",
-				generationName: "ai.streamText.doStream",
-				settings: args.telemetry,
+				async onFinish(event) {
+					const generationOutputs: GenerationOutput[] = [];
+					const generatedTextOutput =
+						generationContext.operationNode.outputs.find(
+							(output) => output.accessor === "generated-text",
+						);
+					if (generatedTextOutput !== undefined) {
+						generationOutputs.push({
+							type: "generated-text",
+							content: event.text,
+							outputId: generatedTextOutput.id,
+						});
+					}
+
+					const tokenUsage = event.usage;
+					let costInfo = null;
+
+					if (tokenUsage) {
+						costInfo = await calculateDisplayCost(
+							operationNode.content.llm.provider,
+							operationNode.content.llm.id,
+							tokenUsage,
+						);
+					}
+
+					const reasoningOutput = generationContext.operationNode.outputs.find(
+						(output) => output.accessor === "reasoning",
+					);
+					if (reasoningOutput !== undefined && event.reasoning !== undefined) {
+						generationOutputs.push({
+							type: "reasoning",
+							content: event.reasoning,
+							outputId: reasoningOutput.id,
+						});
+					}
+					const sourceOutput = generationContext.operationNode.outputs.find(
+						(output) => output.accessor === "source",
+					);
+					if (sourceOutput !== undefined && event.sources.length > 0) {
+						const sources = await Promise.all(
+							event.sources.map(async (source) => {
+								return {
+									sourceType: "url",
+									id: source.id,
+									url: source.url,
+									title: source.title ?? source.url,
+									providerMetadata: source.providerMetadata,
+								} satisfies UrlSource;
+							}),
+						);
+						generationOutputs.push({
+							type: "source",
+							outputId: sourceOutput.id,
+							sources,
+						});
+					}
+					const completedGeneration = {
+						...runningGeneration,
+						status: "completed",
+						completedAt: Date.now(),
+						outputs: generationOutputs,
+						usage: tokenUsage,
+						messages: appendResponseMessages({
+							messages: [
+								{
+									id: "id",
+									role: "user",
+									content: "",
+								},
+							],
+							responseMessages: event.response.messages,
+						}),
+					} satisfies CompletedGeneration;
+
+					await setGeneration(completedGeneration);
+
+					await handleAgentTimeConsumption({
+						workspaceId,
+						generation: completedGeneration,
+						onConsumeAgentTime: args.context.onConsumeAgentTime,
+					});
+
+					// necessary to send telemetry but not explicitly used
+					const langfuse = createLangfuseTracer({
+						workspaceId,
+						runningGeneration,
+						tags: generateTelemetryTags({
+							provider: operationNode.content.llm.provider,
+							languageModel,
+							toolSet: preparedToolSet.toolSet,
+							configurations: operationNode.content.llm.configurations,
+							providerOptions,
+						}),
+						messages: { messages },
+						output: event.text,
+						usage: {
+							input: tokenUsage?.promptTokens ?? 0,
+							output: tokenUsage?.completionTokens ?? 0,
+							total:
+								(tokenUsage?.promptTokens ?? 0) +
+								(tokenUsage?.completionTokens ?? 0),
+							inputCost: costInfo?.inputCostForDisplay ?? 0,
+							outputCost: costInfo?.outputCostForDisplay ?? 0,
+							totalCost: costInfo?.totalCostForDisplay ?? 0,
+							unit: "TOKENS",
+						},
+						completedGeneration,
+						spanName: "ai.streamText",
+						generationName: "ai.streamText.doStream",
+						settings: args.telemetry,
+					});
+					try {
+						await Promise.all([
+							langfuse.shutdownAsync(),
+							...preparedToolSet.cleanupFunctions.map((cleanupFunction) =>
+								cleanupFunction(),
+							),
+						]);
+					} catch (error) {
+						console.error("Cleanup process failed:", error);
+					}
+				},
+				experimental_telemetry: {
+					isEnabled: args.context.telemetry?.isEnabled,
+					metadata: {
+						...args.telemetry?.metadata,
+						tags: [
+							"auto-instrumented",
+							...generateTelemetryTags({
+								provider: operationNode.content.llm.provider,
+								languageModel,
+								toolSet: preparedToolSet.toolSet,
+								configurations: operationNode.content.llm.configurations,
+								providerOptions:
+									operationNode.content.llm.provider === "anthropic"
+										? providerOptions
+										: undefined,
+							}),
+						],
+					},
+				},
 			});
-			try {
-				await Promise.all([
-					langfuse.shutdownAsync(),
-					...preparedToolSet.cleanupFunctions.map((cleanupFunction) =>
-						cleanupFunction(),
-					),
-				]);
-			} catch (error) {
-				console.error("Cleanup process failed:", error);
-			}
-		},
-		experimental_telemetry: {
-			isEnabled: args.context.telemetry?.isEnabled,
-			metadata: {
-				...args.telemetry?.metadata,
-				tags: [
-					"auto-instrumented",
-					...generateTelemetryTags({
-						provider: operationNode.content.llm.provider,
-						languageModel,
-						toolSet: preparedToolSet.toolSet,
-						configurations: operationNode.content.llm.configurations,
-						providerOptions:
-							operationNode.content.llm.provider === "anthropic"
-								? providerOptions
-								: undefined,
-					}),
-				],
-			},
+			return streamTextResult;
 		},
 	});
-	return streamTextResult;
 }
 
 function generationModel(languageModel: TextGenerationLanguageModelData) {
